Search index workaround fix
